refactor: migrate to TS 4
