--- .yarnrc copy.yml
@@ -1,6 +1,0 @@
-yarnPath: .yarn/releases/yarn-3.2.0.cjs
-nodeLinker: node-modules
-
-checksumBehavior: update
-
-defaultSemverRangePrefix: ""
