+++ package.json
@@ -15,9 +15,9 @@
     "@emotion/styled": "~11.10.4",
     "@fortawesome/free-solid-svg-icons": "~6.2.0",
     "@fortawesome/react-fontawesome": "~0.2.0",
-    "@graasp/apps-query-client": "github:graasp/graasp-apps-query-client.git",
+    "@graasp/apps-query-client": "github:graasp/graasp-apps-query-client",
     "@graasp/sdk": "github:graasp/graasp-sdk",
-    "@graasp/ui": "github:graasp/graasp-ui.git",
+    "@graasp/ui": "github:graasp/graasp-ui",
     "@jridgewell/trace-mapping": "~0.3.15",
     "@material-ui/core": "~4.12.4",
     "@material-ui/icons": "~4.11.3",
