+++ src/components/context/AppDataContext.tsx
@@ -1,11 +1,9 @@
 import { List } from 'immutable';
 
 import React, { FC, PropsWithChildren, createContext, useMemo } from 'react';
-
 import { AppData } from '@graasp/apps-query-client';
-
+import { Loader } from '@graasp/ui';
 import { MUTATION_KEYS, hooks, useMutation } from '../../config/queryClient';
-import Loader from '../common/Loader';
 import {
   DeleteAppDataType,
   PatchAppDataType,
