+++ src/components/views/TasksManager.tsx
@@ -1,54 +1,76 @@
-import React from 'react';
-import {
-  DragDropContext,
-  DropResult,
-  ResponderProvided,
-} from 'react-beautiful-dnd';
+import React, { useEffect, useState } from 'react';
+import { DragDropContext, DropResult } from 'react-beautiful-dnd';
 import { Grid } from '@mui/material';
 import { List } from 'immutable';
-import { TASK_LABELS } from '../../config/settings';
 import TasksList from '../main/TasksList';
 import MembersList from '../main/MembersList';
-import { ExistingTaskType, TaskType } from '../../config/appDataTypes';
-import { Member } from '../../types/member';
+import {
+  APP_DATA_TYPES,
+  ExistingTaskType,
+  TaskType,
+} from '../../config/appDataTypes';
+import { useAppDataContext } from '../context/AppDataContext';
+import { useAppActionContext } from '../context/AppActionContext';
+import { useMembersContext } from '../context/MembersContext';
+import { COLORS, TASK_LABELS } from '../../config/constants';
+import { APP_ACTION_TYPES } from '../../config/appActionTypes';
 
-type TasksManagerProps = {
-  tasks: List<ExistingTaskType>;
-  addTask: (task: TaskType) => void;
-  updateTask: (task: ExistingTaskType) => void;
-  deleteTask: (id: string) => void;
-  members: List<Member>;
-};
+const TasksManager = (): JSX.Element => {
+  // get the appData array and a callback to post new appData
+  const { postAppData, patchAppData, deleteAppData, appDataArray } =
+    useAppDataContext();
 
-const TasksManager = (props: TasksManagerProps): JSX.Element => {
-  const { tasks, addTask, updateTask, deleteTask, members } = props;
-  const renderTasksList = (
-    title: string,
-    label: string,
-    add = false,
-  ): JSX.Element => {
-    // eslint-disable-next-line react/destructuring-assignment
-    const tasksArray = tasks
-      .filter(({ data }) => data.label === label)
-      .toArray();
+  const { postAppAction } = useAppActionContext();
 
-    return (
-      <Grid item md={12} lg={4}>
-        <div>
-          <TasksList
-            title={title}
-            label={label}
-            tasks={tasksArray}
-            addComponent={add}
-            addTask={addTask}
-            updateTask={updateTask}
-            deleteTask={deleteTask}
-            // eslint-disable-next-line no-use-before-define
-            members={members}
-          />
-        </div>
-      </Grid>
-    );
+  const [tasks, setTasks] = useState<List<ExistingTaskType>>(List());
+
+  // get the members having access to the space
+  const members = useMembersContext().map((member, index) => ({
+    ...member,
+    color: COLORS[index],
+  }));
+
+  useEffect(() => {
+    const newTasks = appDataArray.filter(
+      ({ type }) => type === APP_DATA_TYPES.TASK,
+    ) as List<ExistingTaskType>;
+    if (newTasks) {
+      setTasks(newTasks);
+    }
+  }, [appDataArray]);
+
+  const addTask = (newTask: TaskType): void => {
+    console.debug('Post app data (newTask): ', newTask);
+    console.debug('postAppData: ', postAppData);
+    postAppData(newTask);
+
+    postAppAction({
+      type: APP_ACTION_TYPES.ADD,
+      data: {
+        ...newTask.data,
+      },
+    });
+  };
+
+  const updateTask = (newTask: ExistingTaskType): void => {
+    patchAppData(newTask);
+    postAppAction({
+      type: APP_ACTION_TYPES.EDIT,
+      data: {
+        ...newTask.data,
+        id: newTask.id,
+      },
+    });
+  };
+
+  const deleteTask = (id: string): void => {
+    deleteAppData({ id });
+    postAppAction({
+      type: APP_ACTION_TYPES.DELETE,
+      data: {
+        id,
+      },
+    });
   };
 
   const handleDragEnd = (result: DropResult): void => {
@@ -79,6 +101,33 @@ const TasksManager = (props: TasksManagerProps): JSX.Element => {
     };
 
     updateTask(newTask);
+  };
+
+  const renderTasksList = (
+    title: string,
+    label: string,
+    add = false,
+  ): JSX.Element => {
+    // eslint-disable-next-line react/destructuring-assignment
+    const tasksArray = tasks.filter(({ data }) => data.label === label);
+
+    return (
+      <Grid item md={12} lg={4}>
+        <div>
+          <TasksList
+            title={title}
+            label={label}
+            tasks={tasksArray}
+            addComponent={add}
+            addTask={addTask}
+            updateTask={updateTask}
+            deleteTask={deleteTask}
+            // eslint-disable-next-line no-use-before-define
+            members={members}
+          />
+        </div>
+      </Grid>
+    );
   };
 
   return (
