--- src/config/actionTypes.js
@@ -1,8 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export const ACTION_TYPES = {
-  SAVE: 'save',
-  ADD: 'add',
-  MOVE: 'move',
-  EDIT: 'edit',
-  DELETE: 'delete',
-};
