--- src/config/contexts.js
@@ -1,6 +1,0 @@
-/* eslint-disable-next-line import/prefer-default-export */
-export const CONTEXTS = {
-  BUILDER: 'builder',
-  PLAYER: 'player',
-  ANALYZER: 'analyzer',
-};
