+++ src/config/settings.js
@@ -1,5 +1,4 @@
-import { CONTEXTS } from './contexts';
-import { REACT_APP_MOCK_API } from './env';
+import { CONTEXTS } from './constants';
 
 export const DEFAULT_LANG = 'en';
 
@@ -14,12 +13,6 @@ export const PERMISSION_LEVELS = {
 
 export const DEFAULT_PERMISSION = 'read';
 
-export const TASK_LABELS = {
-  TODO: 'todo',
-  IN_PROGRESS: 'in_progress',
-  COMPLETED: 'completed',
-};
-
 export const MAX_NUM_FILES = 1;
 export const MAX_FILE_SIZE = 5 * 1000 * 1000;
 
@@ -29,5 +22,3 @@ export const DEFAULT_LOCAL_CONTEXT = {
   context: CONTEXTS.PLAYER,
   apiHost: DEFAULT_API_HOST,
 };
-
-export const MOCK_API = REACT_APP_MOCK_API === 'true';
